revocation: place remaining functions in `crl`.
